requirements.txt
pandas
yahooquery
flask
apscheduler
# lxml==5.3.2
psycopg2-binary